requirements.txt
# Core dependencies
requests
python-dotenv

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Optional: Development tools
# pytest>=7.4.0
# httpx>=0.25.0  # For testing FastAPI endpoints
